requirements.txt
pyinstaller==6.3.0
pillow==10.2.0


